+++ jwp-basic-gradle/src/main/java/next/dao/UserDao.java
@@ -1,59 +1,16 @@
 package next.dao;
 
-import core.jdbc.JdbcTemplate;
 import java.util.List;
 import next.model.User;
 
-public class UserDao {
+public interface UserDao {
 
-    private static final JdbcTemplate jdbcTemplate = JdbcTemplate.getInstance();
+    void insert(User user);
 
-    public void insert(User user) {
-        jdbcTemplate.update(
-                "INSERT INTO USERS VALUES (?, ?, ?, ?)",
-                preparedStatement -> {
-                    preparedStatement.setString(1, user.getUserId());
-                    preparedStatement.setString(2, user.getPassword());
-                    preparedStatement.setString(3, user.getName());
-                    preparedStatement.setString(4, user.getEmail());
-                }
-        );
-    }
+    void update(final User user);
 
-    public void update(final User user) {
-        jdbcTemplate.update(
-                "UPDATE USERS SET password = ?, name = ?, email = ? WHERE userId = ?",
-                preparedStatement -> {
-                    preparedStatement.setString(1, user.getPassword());
-                    preparedStatement.setString(2, user.getName());
-                    preparedStatement.setString(3, user.getEmail());
-                    preparedStatement.setString(4, user.getUserId());
-                }
-        );
-    }
+    User findById(String userId);
 
-    public User findById(String userId) {
-        return jdbcTemplate.queryForObject(
-                "SELECT userId, password, name, email FROM USERS WHERE userId = ?",
-                resultSet -> new User(
-                        resultSet.getString("userId"),
-                        resultSet.getString("password"),
-                        resultSet.getString("name"),
-                        resultSet.getString("email")),
-                userId
-        );
-    }
-
-    public List<User> findAll() {
-        return jdbcTemplate.query(
-                "SELECT userId, password, name, email from USERS",
-                resultSet -> new User(
-                        resultSet.getString("userId"),
-                        resultSet.getString("password"),
-                        resultSet.getString("name"),
-                        resultSet.getString("email")
-                )
-        );
-    }
+    List<User> findAll();
 
 }
