+++ jwp-basic-gradle/src/test/java/next/dao/UserDaoTest.java
@@ -3,15 +3,10 @@ package next.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import core.jdbc.ConnectionManager;
 import java.util.List;
 import next.model.User;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.jdbc.datasource.init.DatabasePopulatorUtils;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 
 public class UserDaoTest extends DaoTest {
 
@@ -19,7 +14,7 @@ public class UserDaoTest extends DaoTest {
     @Test
     public void crud() {
         User expected = new User("userId", "password", "name", "javajigi@email.com");
-        UserDao userDao = new UserDao();
+        JdbcUserDao userDao = new JdbcUserDao();
         userDao.insert(expected);
 
         User actual = userDao.findById(expected.getUserId());
@@ -30,7 +25,7 @@ public class UserDaoTest extends DaoTest {
     void findAll() {
         // given
         User expected = new User("userId", "password", "name", "test@email.com");
-        UserDao userDao = new UserDao();
+        JdbcUserDao userDao = new JdbcUserDao();
 
         // when
         userDao.insert(expected);
